Jest Action: Only run tests related to JS files
This ensures that we don't run Jest if there are no changes to JS files. Additionally, it will make it so tha we only run test files that are related to changes that were made. If a test file was changed then that will be run - and if anon-test file was changed then it, and anything that imports it, will have its test files run.

Issue: none

Test plan:
Will this have a negative impact on the mobile repo? Are they expecting to run all tests all the time? It will definitely help to improve performance!